+++ src/FunFair.BuildCheck.ProjectChecks/Settings/ProjectsShouldHaveTrimAnalyzerConfiguredPolicy.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using FunFair.BuildCheck.Interfaces;
@@ -10,6 +11,7 @@ namespace FunFair.BuildCheck.ProjectChecks.Settings;
 public sealed class ProjectsShouldHaveTrimAnalyzerConfiguredPolicy : IProjectCheck
 {
     private const string SETTING = "EnableTrimAnalyzer";
+    private const string IS_TRIMMABLE = "IsTrimmable";
 
     private readonly ILogger<ProjectsShouldHaveTrimAnalyzerConfiguredPolicy> _logger;
 
@@ -22,9 +24,30 @@ public sealed class ProjectsShouldHaveTrimAnalyzerConfiguredPolicy : IProjectChe
 
     public ValueTask CheckAsync(ProjectContext project, CancellationToken cancellationToken)
     {
-        if (!project.HasProperty(SETTING))
+        string? trimAnalyzer = project.GetProperty(SETTING);
+
+        if (string.IsNullOrWhiteSpace(trimAnalyzer))
         {
             this._logger.ProjectShouldConfigureTrimAnalyzer(projectName: project.Name, property: SETTING);
+
+            return ValueTask.CompletedTask;
+        }
+
+        bool isTrimAnalyzerTrue = StringComparer.OrdinalIgnoreCase.Equals(x: trimAnalyzer, y: "true");
+        bool isTrimAnalyzerFalse = StringComparer.OrdinalIgnoreCase.Equals(x: trimAnalyzer, y: "false");
+
+        if (!isTrimAnalyzerTrue && !isTrimAnalyzerFalse)
+        {
+            this._logger.ProjectTrimAnalyzerMustBeTrueOrFalse(projectName: project.Name, property: SETTING, value: trimAnalyzer);
+
+            return ValueTask.CompletedTask;
+        }
+
+        string? isTrimmable = project.GetProperty(IS_TRIMMABLE);
+
+        if (StringComparer.OrdinalIgnoreCase.Equals(x: isTrimmable, y: "true") && !isTrimAnalyzerTrue)
+        {
+            this._logger.ProjectMustEnableTrimAnalyzerWhenTrimmable(projectName: project.Name, property: SETTING);
         }
 
         return ValueTask.CompletedTask;